+++ crates/iroha_core/src/kura.rs
@@ -276,6 +276,7 @@ impl Kura {
                     error!(?error, "Failed to store block");
                     panic!("Kura has encountered a fatal IO error.");
                 }
+                latest_written_block_hash = Some(block.hash());
             }
         }
     }
@@ -888,7 +889,7 @@ mod tests {
         assert_eq!(block_store.read_block_index(2).unwrap(), (6, 3));
         assert_eq!(block_store.read_block_index(3).unwrap(), (1, 2));
 
-        // or equivilant
+        // or equivalent
         {
             let should_be = indices([(2, 9), (0, 0), (6, 3), (1, 2)]);
             let mut is = indices([(0, 0), (0, 0), (0, 0), (0, 0)]);