+++ crates/iroha_test_network/src/lib.rs
@@ -42,7 +42,7 @@ use tokio::{
     io::{AsyncBufReadExt, AsyncWriteExt, BufReader},
     process::Child,
     runtime::{self, Runtime},
-    sync::{broadcast, oneshot, watch, Mutex},
+    sync::{broadcast, oneshot, watch, Barrier, Mutex},
     task::{spawn_blocking, JoinSet},
     time::timeout,
 };
@@ -210,7 +210,7 @@ impl Network {
         self.peers.iter().map(|x| x.id.clone()).collect()
     }
 
-    /// Resolves when all _running_ peers have at least N blocks
+    /// Resolves when all _running_ peers have at least N non-empty blocks
     /// # Errors
     /// If this doesn't happen within a timeout.
     pub async fn ensure_blocks(&self, height: u64) -> Result<&Self> {
@@ -349,7 +349,8 @@ impl NetworkBuilder {
                 ))),
             ]
             .into_iter()
-            .chain(self.extra_isi),
+            .chain(self.extra_isi)
+            .chain(extra_isi),
             topology,
         );
 
@@ -465,6 +466,7 @@ pub struct NetworkPeer {
     is_running: Arc<AtomicBool>,
     events: broadcast::Sender<PeerLifecycleEvent>,
     block_height: watch::Sender<Option<u64>>,
+    block_height_non_empty: watch::Sender<Option<u64>>,
     // dropping these the last
     port_p2p: Arc<AllocatedPort>,
     port_api: Arc<AllocatedPort>,
@@ -492,6 +494,7 @@ impl NetworkPeer {
 
         let (events, _rx) = broadcast::channel(32);
         let (block_height, _rx) = watch::channel(None);
+        let (block_height_non_empty, _rx) = watch::channel(None);
 
         let result = Self {
             id,
@@ -502,6 +505,7 @@ impl NetworkPeer {
             is_running: Default::default(),
             events,
             block_height,
+            block_height_non_empty,
             port_p2p: Arc::new(port_p2p),
             port_api: Arc::new(port_api),
         };
@@ -523,9 +527,10 @@ impl NetworkPeer {
     ///
     /// Passed configuration must contain network topology in the `trusted_peers` parameter.
     ///
-    /// This function doesn't wait for peer server to start working, or for it to commit genesis block.
-    /// Iroha could as well terminate immediately with an error, and it is not tracked by this function.
-    /// Use [`Self::events`]/[`Self::once`] to monitor peer's lifecycle.
+    /// This function waits for peer server to start working,
+    /// in particular it waits for `/status` response and connects to event stream.
+    /// However it doesn't wait for genesis block to be commited.
+    /// See [`Self::events`]/[`Self::once`]/[`Self::once_block`] to monitor peer's lifecycle.
     ///
     /// # Panics
     /// If peer was not started.
@@ -554,7 +559,7 @@ impl NetworkPeer {
                 format!("127.0.0.1:{}", self.port_p2p),
             )
             .write(["torii", "address"], format!("127.0.0.1:{}", self.port_api))
-            .write(["logger", "format"], "json");
+            .write(["logger", "format"], "pretty");
 
         let config_path = self.dir.path().join(format!("run-{run_num}-config.toml"));
         let genesis_path = self.dir.path().join(format!("run-{run_num}-genesis.scale"));
@@ -596,7 +601,8 @@ impl NetworkPeer {
                 .unwrap();
             tasks.spawn(async move {
                 let mut lines = BufReader::new(output).lines();
-                while let Ok(Some(line)) = lines.next_line().await {
+                while let Ok(Some(mut line)) = lines.next_line().await {
+                    line.push('\n');
                     file.write_all(line.as_bytes())
                         .await
                         .expect("writing logs to file shouldn't fail");
@@ -635,6 +641,7 @@ impl NetworkPeer {
             is_running: self.is_running.clone(),
             events: self.events.clone(),
             block_height: self.block_height.clone(),
+            block_height_non_empty: self.block_height_non_empty.clone(),
         };
         tasks.spawn(async move {
             if let Err(err) = peer_exit.monitor(shutdown_rx).await {
@@ -643,11 +650,15 @@ impl NetworkPeer {
             }
         });
 
+        let barrier = Arc::new(Barrier::new(2));
+
         {
             let log_prefix = log_prefix.clone();
             let client = self.client();
             let events_tx = self.events.clone();
             let block_height_tx = self.block_height.clone();
+            let block_height_non_empty_tx = self.block_height_non_empty.clone();
+            let barrier = barrier.clone();
             tasks.spawn(async move {
                 let status_client = client.clone();
                 let status = backoff::future::retry(
@@ -669,7 +680,8 @@ impl NetworkPeer {
                 .await
                 .expect("there is no max elapsed time");
                 let _ = events_tx.send(PeerLifecycleEvent::ServerStarted);
-                let _ = block_height_tx.send(Some(status.blocks));
+                let _ = block_height_tx.send_replace(Some(status.blocks));
+                let _ = block_height_non_empty_tx.send_replace(Some(status.blocks_non_empty));
                 eprintln!("{log_prefix} server started, {status:?}");
 
                 let mut events = client
@@ -680,19 +692,34 @@ impl NetworkPeer {
                         panic!("cannot proceed")
                     });
 
+                // We need to connect to event stream before [start] function returns.
+                // Otherwise there can be such situation:
+                // * [start] function returns
+                // * test code submits some transactions and peer applied some blocks
+                // * we do not yet connect to event stream,
+                //   so `height` and `height_non_empty` will not be updated
+                // * test code uses `ensure_blocks` function which relies on `height_non_empty`
+                barrier.wait().await;
+
                 while let Some(Ok(event)) = events.next().await {
                     if let EventBox::Pipeline(PipelineEventBox::Block(block)) = event {
                         if *block.status() == BlockStatus::Applied {
                             let height = block.header().height().get();
-                            eprintln!("{log_prefix} BlockStatus::Applied height={height}",);
+                            let is_empty = block.header().transactions_hash().is_none();
+                            eprintln!("{log_prefix} BlockStatus::Applied height={height}, is_empty={is_empty}");
                             let _ = events_tx.send(PeerLifecycleEvent::BlockApplied { height });
                             block_height_tx.send_modify(|x| *x = Some(height));
+
+                            let status = client.get_status().expect("Can't get status");
+                            block_height_non_empty_tx.send_modify(|x| *x = Some(status.blocks_non_empty));
                         }
                     }
                 }
                 eprintln!("{log_prefix} events stream is closed");
             });
         }
+
+        barrier.wait().await;
 
         *run_guard = Some(PeerRun {
             tasks,
@@ -755,11 +782,11 @@ impl NetworkPeer {
         }
     }
 
-    /// Wait until peer's block height reaches N.
+    /// Wait until peer's count of non-empty blocks reaches N.
     ///
-    /// Resolves immediately if peer is already running _and_ its current block height is greater or equal to N.
+    /// Resolves immediately if peer is already running _and_ has at least N non-empty blocks committed.
     pub async fn once_block(&self, n: u64) {
-        let mut recv = self.block_height.subscribe();
+        let mut recv = self.block_height_non_empty.subscribe();
 
         if recv.borrow().map(|x| x >= n).unwrap_or(false) {
             return;
@@ -824,10 +851,6 @@ impl NetworkPeer {
             .await
             .expect("should not panic")
     }
-
-    pub fn blocks(&self) -> watch::Receiver<Option<u64>> {
-        self.block_height.subscribe()
-    }
 }
 
 /// Compare by ID
@@ -862,6 +885,7 @@ struct PeerExit {
     is_running: Arc<AtomicBool>,
     events: broadcast::Sender<PeerLifecycleEvent>,
     block_height: watch::Sender<Option<u64>>,
+    block_height_non_empty: watch::Sender<Option<u64>>,
 }
 
 impl PeerExit {
@@ -875,6 +899,7 @@ impl PeerExit {
         let _ = self.events.send(PeerLifecycleEvent::Terminated { status });
         self.is_running.store(false, Ordering::Relaxed);
         self.block_height.send_modify(|x| *x = None);
+        self.block_height_non_empty.send_modify(|x| *x = None);
 
         Ok(())
     }